Safe user deletion logic implemented

+++ src/main/java/com/flicko/TaskMan/services/UserService.java
@@ -2,10 +2,14 @@ package com.flicko.TaskMan.services;
 
 import com.flicko.TaskMan.DTOs.UserRoleUpdate;
 import com.flicko.TaskMan.DTOs.UserUpdate;
+import com.flicko.TaskMan.enums.UserRole;
+import com.flicko.TaskMan.models.Task;
 import com.flicko.TaskMan.models.User;
+import com.flicko.TaskMan.repos.TaskRepository;
 import com.flicko.TaskMan.repos.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -14,6 +18,8 @@ import java.util.List;
 public class UserService {
 
     private final UserRepository userRepository;
+
+    private final TaskRepository taskRepository;
 
     public List<User> getAllUsers() {
         return userRepository.findAll();
@@ -28,11 +34,22 @@ public class UserService {
         return userRepository.save(user);
     }
 
+    @Transactional
     public void deleteUser(Long id) {
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("User not found"));
 
+        if (user.getRole() == UserRole.ADMIN && userRepository.countByRole(UserRole.ADMIN) <= 1){
+            throw new RuntimeException("Can't delete only ADMIN");
+        }
+
+        List<Task> tasks = taskRepository.findByUserId(id);
+
+        tasks.forEach(t -> t.setUser(null));
+        taskRepository.saveAll(tasks);
+
         userRepository.delete(user);
+
     }
 
     public User updateUser(Long id, UserUpdate user) {
